requirements.txt
pydantic~=1.8.2
fastapi~=0.68.2
requests~=2.26.0
starlette~=0.14.2
strawberry-graphql~=0.81.0
uvicorn~=0.15.0
